initial refactoring
