+++ requirements.txt
@@ -1,0 +1,4 @@
+torch==2.1.0
+transformers==4.36.1
+multiaddr @ git+https://github.com/mobiusml/hqq.git@37502bea31f2969c6680c0c4a88ca74b3bb234a5
+numpy==1.23.5
